plugins/cat_inventory_management/requirements.txt
pandas
psycopg2-binary
tabulate